requirements.txt
discord.py
requests
pydub
aiohttp
yt-dlp
python-dotenv
PyNaCl